+++ R/visualization.R
@@ -4,7 +4,7 @@
 #' @description Create density distribution of coverage summary statistics.
 #' Use the coverage summary file created with coverage_summary function.
 #' @param data Coverage summary file.
-#' @param aes.colour GGPLOT2 aesthetics colour, 
+#' @param aes.colour GGPLOT2 aesthetics colour,
 #' e.g. aes(y = ..scaled.., color = COVERAGE_GROUP).
 #' @param adjust.bin Adjust GGPLOT2 bin size (0 to 1).
 #' @export
@@ -14,23 +14,23 @@
 #' @import readr
 
 plot_density_distribution_coverage <- function(data, aes.colour, adjust.bin) {
-  
+
   VALUE <- NULL
 
-  
-  
-  ggplot(data, aes(x = VALUE, na.rm = T))+
-    geom_line(aes.colour, stat = "density", size = 0.5, adjust = adjust.bin)+
-    labs(x = "Depth of coverage(read number)")+
-    labs(y = "Loci (scaled density)")+
-    expand_limits(x = 0)+
-    theme(
+
+
+  ggplot2::ggplot(data, ggplot2::aes(x = VALUE, na.rm = TRUE)) +
+    ggplot2::geom_line(aes.colour, stat = "density", size = 0.5, adjust = adjust.bin) +
+    ggplot2::labs(x = "Depth of coverage(read number)") +
+    ggplot2::labs(y = "Loci (scaled density)") +
+    ggplot2::expand_limits(x = 0) +
+    ggplot2::theme(
       legend.position = "none",
-      axis.title.x = element_text(size = 10, family = "Helvetica", face = "bold"),
-      axis.title.y = element_text(size = 10, family = "Helvetica", face = "bold"),
-      legend.title = element_text(size = 10, family = "Helvetica", face = "bold"),
-      legend.text = element_text(size = 10, family = "Helvetica", face = "bold"),
-      strip.text.x = element_text(size = 10, family = "Helvetica", face = "bold"))
+      axis.title.x = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      axis.title.y = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      legend.title = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      legend.text = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      strip.text.x = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"))
 }
 
 
@@ -44,27 +44,27 @@ plot_density_distribution_coverage <- function(data, aes.colour, adjust.bin) {
 #' @rdname plot_boxplot_coverage
 
 plot_boxplot_coverage <- function(data) {
-  
-  POP_ID <- NULL
-  VALUE <- NULL
-  POP_ID <- NULL
-  VALUE <- NULL
-  
 
-  ggplot(data, aes(x = factor(POP_ID), y = VALUE, na.rm = T))+
-    geom_violin(trim = F)+
-    geom_boxplot(width = 0.1, fill = "black", outlier.colour = NA)+
-    stat_summary(fun.y = "mean", geom = "point", shape = 21, size = 2.5, fill = "white")+
-    labs(x = "Sampling sites")+
-    labs(y = "Read depth coverage")+
-    facet_wrap(facets = ~COVERAGE_GROUP, scales = "free")+
-    theme(
+  POP_ID <- NULL
+  VALUE <- NULL
+  POP_ID <- NULL
+  VALUE <- NULL
+
+
+  ggplot2::ggplot(data, ggplot2::aes(x = factor(POP_ID), y = VALUE, na.rm = TRUE)) +
+    ggplot2::geom_violin(trim = FALSE) +
+    ggplot2::geom_boxplot(width = 0.1, fill = "black", outlier.colour = NA) +
+    ggplot2::stat_summary(fun.y = "mean", geom = "point", shape = 21, size = 2.5, fill = "white") +
+    ggplot2::labs(x = "Sampling sites") +
+    ggplot2::labs(y = "Read depth coverage") +
+    ggplot2::facet_wrap(facets = ~COVERAGE_GROUP, scales = "free") +
+    ggplot2::theme(
       legend.position = "none",
-      axis.title.x = element_text(size = 10, family = "Helvetica", face = "bold"),
-      axis.title.y = element_text(size = 10, family = "Helvetica", face = "bold"),
-      legend.title = element_text(size = 10, family = "Helvetica", face = "bold"),
-      legend.text = element_text(size = 10, family = "Helvetica", face = "bold"),
-      strip.text.x = element_text(size = 10, family = "Helvetica", face = "bold")
+      axis.title.x = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      axis.title.y = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      legend.title = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      legend.text = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      strip.text.x = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold")
     )
 }
 
@@ -73,38 +73,38 @@ plot_boxplot_coverage <- function(data) {
 
 #' @title Visual diagnostic of coverage imbalance
 #' @description GBS data and STACKS pipeline sometimes output REF and ALT
-#' alleles that have coverage imbalance, i.e. the coverage is not equal 
-#' and skewed towards the REF or ALT alleles. 
+#' alleles that have coverage imbalance, i.e. the coverage is not equal
+#' and skewed towards the REF or ALT alleles.
 #' Thw density distribution figure of coverage imbalance between REF and ALT
 #' alleles will highlight the problem in you data.
 #' @param tidy.vcf.file The tidy VCF file created with tidy_genomic_data.
 #' @param pop.levels Character string defining your ordered populations.
 #' @param read.depth.threshold Define the threshold you wish to analyse.
-#' @param aes.colour GGPLOT2 aesthetics, 
+#' @param aes.colour GGPLOT2 aesthetics,
 #' e.g. aes(y = ..count..).
 #' @param adjust.bin Adjust GGPLOT2 bin size (0 to 1).
 #' @return 4-plots highlighting the different combination of under
-#'  or over the coverage threshold and mean genotype likelihood. 
+#'  or over the coverage threshold and mean genotype likelihood.
 #'  Y- axis show the distribution of genotypes.
-#'  X- axis show the coverage imbalance the 
+#'  X- axis show the coverage imbalance the
 #'  Negative ratio (left x axis) : REF > ALT.
 #'  Positive ratio (right x axis) : ALT > REF.
 #' @details The figures shows the results of the of coverage threshold
-#' selected and mean genotype likelihood. 
+#' selected and mean genotype likelihood.
 #' You can test different threshold to inspect your data.
 #' Ideally the lower left pannel of the 4-plot should be empty. If it is, this
 #' shows that setting the threshold of the genotype likelihood filter
 #' to the mean or close to it take care of the allelic coverage imbalance.
 #' #' e.g. fig <- plot_coverage_imbalance_diagnostic(
 #' tidy.vcf.file, pop.levels, read.depth.threshold, aes.colour, adjust.bin)
-#' Use ( fig + facet_grid(GROUP_GL ~ GROUP_COVERAGE)). The ratio is calculated : 
+#' Use ( fig + facet_grid(GROUP_GL ~ GROUP_COVERAGE)). The ratio is calculated :
 #' (read depth ALT allele - read depth REF allele)/(read depth ALT allele + read depth REF allele).
 #' @export
 #' @rdname plot_coverage_imbalance_diagnostic
 
 
 plot_coverage_imbalance_diagnostic <- function(tidy.vcf.file, pop.levels, read.depth.threshold, aes.colour, adjust.bin) {
-  
+
   INDIVIDUALS <- NULL
   POP_ID <- NULL
   READ_DEPTH <- NULL
@@ -112,30 +112,30 @@ plot_coverage_imbalance_diagnostic <- function(tidy.vcf.file, pop.levels, read.d
   GROUP_COVERAGE <- NULL
   GROUP_GL <- NULL
   ALLELE_COVERAGE_RATIO <- NULL
-  
-  
-  
-  if (is.vector(tidy.vcf.file) == "TRUE") {
-    data <- read_tsv(tidy.vcf.file, col_names = T, col_types = "diidccddccccdddddc") %>%
-      mutate(INDIVIDUALS = factor(INDIVIDUALS))
+
+
+
+  if (is.vector(tidy.vcf.file)) {
+    data <- readr::read_tsv(tidy.vcf.file, col_names = T, col_types = "diidccddccccdddddc") %>%
+      dplyr::mutate(INDIVIDUALS = factor(INDIVIDUALS))
   } else {
     data <- tidy.vcf.file
   }
-  
-  if (missing(pop.levels) == "TRUE") {
+
+  if (missing(pop.levels)) {
     data <- data
   } else {
     data <- suppressWarnings(
       data %>%
-      mutate(POP_ID = factor(POP_ID, levels = pop.levels, ordered = T))
+      dplyr::mutate(POP_ID = factor(POP_ID, levels = pop.levels, ordered = T))
     )
   }
-  
-  below.threshold <- stri_join("READ DEPTH <=", read.depth.threshold, sep = " ")
-  over.threshold <- stri_join("READ DEPTH >", read.depth.threshold, sep = " ")
-  
+
+  below.threshold <- stringi::stri_join("READ DEPTH <=", read.depth.threshold, sep = " ")
+  over.threshold <- stringi::stri_join("READ DEPTH >", read.depth.threshold, sep = " ")
+
   imbalance.coverage <- data %>%
-    mutate(
+    dplyr::mutate(
       #       GROUP_COVERAGE = ifelse(READ_DEPTH <= read.depth.threshold, "READ DEPTH <= 8", "READ DEPTH > 8"),
       GROUP_COVERAGE = ifelse(READ_DEPTH <= read.depth.threshold, below.threshold, over.threshold),
       GROUP_GL = ifelse(GL < (mean(data$GL, na.rm = T)), "< mean GL", ">= mean GL"),
@@ -143,58 +143,58 @@ plot_coverage_imbalance_diagnostic <- function(tidy.vcf.file, pop.levels, read.d
       #       GROUP_COVERAGE = factor(GROUP_COVERAGE, levels = c("READ DEPTH <= 8", "READ DEPTH > 8"), ordered = T),
       GROUP_GL = factor(GROUP_GL, levels = c("< mean GL", ">= mean GL"), ordered = T)
     ) %>%
-    filter(GROUP_COVERAGE != "NA" & GROUP_GL != "NA")
-  
-  ggplot(imbalance.coverage, aes(x = ALLELE_COVERAGE_RATIO, na.rm = T))+
-    #   geom_bar()+
-    geom_line(aes.colour, stat = "density", adjust = adjust.bin)+
-    labs(x="Coverage imbalance between REF and ALT alleles (ratio)")+
-    labs(y="Distribution of genotypes")+
-    theme(
-      axis.title.x = element_text(size = 12, family = "Helvetica", face = "bold"),
-      axis.title.y = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      legend.title = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      legend.text = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      strip.text.x = element_text(size = 12, family = "Helvetica", face = "bold"),
-      strip.text.y = element_text(size = 12, family = "Helvetica", face = "bold")
+    dplyr::filter(GROUP_COVERAGE != "NA" & GROUP_GL != "NA")
+
+  ggplot2::ggplot(imbalance.coverage, ggplot2::aes(x = ALLELE_COVERAGE_RATIO, na.rm = T)) +
+    #   geom_bar() +
+    ggplot2::geom_line(aes.colour, stat = "density", adjust = adjust.bin) +
+    ggplot2::labs(x="Coverage imbalance between REF and ALT alleles (ratio)") +
+    ggplot2::labs(y="Distribution of genotypes") +
+    ggplot2::theme(
+      axis.title.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      axis.title.y = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.title = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.text = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      strip.text.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      strip.text.y = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold")
     )
-  
+
 }
 
 
 #' @title Figure density distribution of genotype likelihood summary statistics
-#' @description Create density distribution of genotype likelihood 
+#' @description Create density distribution of genotype likelihood
 #' summary statistics.
-#' Use the long version of coverage summary file created with 
+#' Use the long version of coverage summary file created with
 #' genotype_likelihood_summary function ($gl.summary.long).
 #' @param data Genotype likelihood summary file.
-#' @param aes.colour GGPLOT2 aesthetics colour, 
+#' @param aes.colour GGPLOT2 aesthetics colour,
 #' e.g. aes(y = ..scaled.., color = GENOTYPE_LIKELIHOOD_GROUP).
 #' @param adjust.bin Adjust GGPLOT2 bin size (0 to 1).
 #' @export
 #' @rdname plot_density_distribution_genotype_likelihood
 #' @seealso \link{summary_genotype_likelihood}
 
-plot_density_distribution_genotype_likelihood <- function(data, aes.colour, 
+plot_density_distribution_genotype_likelihood <- function(data, aes.colour,
                                                             adjust.bin) {
-  
+
   VALUE <- NULL
 
   # BREAKS <- seq(0, 150, by = 20)
-  suppressWarnings(ggplot(data, aes(x = VALUE, na.rm = T))+
-    geom_line(aes.colour, stat = "density", size = 0.5, adjust = adjust.bin)+
-    #  scale_colour_manual(name = "Filters", values = c("black", "blue"))+
-    #  scale_x_continuous(breaks = BREAKS)+
-    labs(x = "Markers Genotype Likelihood")+
-    labs(y = "Markers (scaled density)")+
-    expand_limits(x = 0)+
-    theme(
+  suppressWarnings(ggplot2::ggplot(data, ggplot2::aes(x = VALUE, na.rm = T)) +
+                     ggplot2::geom_line(aes.colour, stat = "density", size = 0.5, adjust = adjust.bin) +
+    #  scale_colour_manual(name = "Filters", values = c("black", "blue")) +
+    #  scale_x_continuous(breaks = BREAKS) +
+      ggplot2::labs(x = "Markers Genotype Likelihood") +
+      ggplot2::labs(y = "Markers (scaled density)") +
+      ggplot2::expand_limits(x = 0) +
+      ggplot2::theme(
       legend.position = "none",
-      axis.title.x = element_text(size = 10, family = "Helvetica", face = "bold"),
-      axis.title.y = element_text(size = 10, family = "Helvetica", face = "bold"),
-      legend.title = element_text(size = 10, family = "Helvetica", face = "bold"),
-      legend.text = element_text(size = 10, family = "Helvetica", face = "bold"),
-      strip.text.x = element_text(size = 10, family = "Helvetica", face = "bold"))
+      axis.title.x = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      axis.title.y = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      legend.title = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      legend.text = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      strip.text.x = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"))
   )
 }
 
@@ -211,38 +211,38 @@ plot_density_distribution_genotype_likelihood <- function(data, aes.colour,
 #' @rdname plot_boxplot_genotype_likelihood
 
 plot_boxplot_genotype_likelihood <- function(data) {
-  
+
   POP_ID <- NULL
   VALUE <- NULL
-  
-  
-  
-  ggplot(data, aes(x = factor(POP_ID), y = VALUE, na.rm = T))+
-    geom_violin(trim = TRUE)+
-    geom_boxplot(width = 0.1, fill = "black", outlier.colour = NA)+
-    stat_summary(fun.y = "mean", geom = "point", shape = 21, size = 2.5, fill = "white")+
-    labs(x = "Sampling sites")+
-    labs(y = "Genotype likelihood of markers")+
-    theme(
+
+
+
+  ggplot2::ggplot(data, ggplot2::aes(x = factor(POP_ID), y = VALUE, na.rm = T)) +
+    ggplot2::geom_violin(trim = TRUE) +
+    ggplot2::geom_boxplot(width = 0.1, fill = "black", outlier.colour = NA) +
+    ggplot2::stat_summary(fun.y = "mean", geom = "point", shape = 21, size = 2.5, fill = "white") +
+    ggplot2::labs(x = "Sampling sites") +
+    ggplot2::labs(y = "Genotype likelihood of markers") +
+    ggplot2::theme(
       legend.position = "none",
-      axis.title.x = element_text(size = 10, family = "Helvetica", face = "bold"),
-      axis.title.y = element_text(size = 10, family = "Helvetica", face = "bold"),
-      axis.text.x = element_text(size = 8, family = "Helvetica", angle = 90, hjust = 1, vjust = 0.5), 
-      legend.title = element_text(size = 10, family = "Helvetica", face = "bold"),
-      legend.text = element_text(size = 10, family = "Helvetica", face = "bold"),
-      strip.text.x = element_text(size = 10, family = "Helvetica", face = "bold")
+      axis.title.x = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      axis.title.y = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      axis.text.x = ggplot2::element_text(size = 8, family = "Helvetica", angle = 90, hjust = 1, vjust = 0.5),
+      legend.title = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      legend.text = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold"),
+      strip.text.x = ggplot2::element_text(size = 10, family = "Helvetica", face = "bold")
     )
 }
 
 
 
 
-#' @title Figure density distribution of minor allele frequency (MAF) 
+#' @title Figure density distribution of minor allele frequency (MAF)
 #' summary statistics.
 #' @description Create density distribution of MAF summary statistics.
 #' @param data sumstats or tidy vcf summarised files.
 #' @param maf.group The GGPLOT2 aes (e.g. aes(x = FREQ_ALT, na.rm = F)).
-#' @param aes.colour GGPLOT2 aesthetics colour, 
+#' @param aes.colour GGPLOT2 aesthetics colour,
 #' e.g. aes(y = ..scaled.., color = POP_ID) or aes(y = ..count.., color = POP_ID)
 #' @param adjust.bin Adjust GGPLOT2 bin size (0 to 1).
 #' @param x.title Title of the x-axis. e.g. "MAF distribution"
@@ -255,30 +255,30 @@ plot_boxplot_genotype_likelihood <- function(data) {
 #' @author Thierry Gosselin \email{thierrygosselin@@icloud.com}
 
 plot_density_distribution_maf <- function(data, maf.group, aes.colour = aes(y = ..scaled.., color = POP_ID), adjust.bin = 1, x.title) {
-  
+
   ..scaled.. <- NULL #get rid of R CMD check note
-  
+
   if (is.vector(data)) {
     data <- read_tsv(data, col_names = T)
   } else {
     data <- data
   }
-  #   font.group <- 
-  graph <- ggplot(data, maf.group)+
-    geom_line(aes.colour, stat = "density", adjust = adjust.bin)+ # pop colored
-    #   scale_colour_manual(name ="Sampling sites", values = colour_palette_sites.pink)+
-    scale_x_continuous(breaks=c(0, 0.05, 0.1, 0.2, 0.5, 1),
-                       labels = c("0", "0.05", "0.1", "0.2", "0.5", "1.0"))+
-    labs(x = x.title)+
-    labs(y = "Density of SNP (scaled)")+
-    expand_limits(y = 0)+
-    theme(
-      axis.title.x = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      axis.title.y = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      legend.title = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      legend.text = element_text(size = 12, family = "Helvetica", face = "bold"),
-      strip.text.y = element_text(angle = 0, size = 12, family = "Helvetica", face = "bold"), 
-      strip.text.x = element_text(size = 12, family = "Helvetica", face = "bold")
+  #   font.group <-
+  graph <- ggplot2::ggplot(data, maf.group) +
+    ggplot2::geom_line(aes.colour, stat = "density", adjust = adjust.bin) + # pop colored
+    #   scale_colour_manual(name ="Sampling sites", values = colour_palette_sites.pink) +
+    ggplot2::scale_x_continuous(breaks=c(0, 0.05, 0.1, 0.2, 0.5, 1),
+                       labels = c("0", "0.05", "0.1", "0.2", "0.5", "1.0")) +
+    ggplot2::labs(x = x.title) +
+    ggplot2::labs(y = "Density of SNP (scaled)") +
+    ggplot2::expand_limits(y = 0) +
+    ggplot2::theme(
+      axis.title.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      axis.title.y = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.title = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.text = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      strip.text.y = ggplot2::element_text(angle = 0, size = 12, family = "Helvetica", face = "bold"),
+      strip.text.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold")
     )
 }
 
@@ -294,7 +294,7 @@ plot_density_distribution_maf <- function(data, maf.group, aes.colour = aes(y = 
 #' @param data sumstats or tidy vcf files.
 #' @param pop.levels Character string defining your ordered populations.
 #' @param het.group = aes(x = HET_MAX, na.rm = F)
-#' @param aes.colour GGPLOT2 aesthetics colour, 
+#' @param aes.colour GGPLOT2 aesthetics colour,
 #' e.g. aes(y = ..scaled.., color = GROUP).
 #' @param adjust.bin Adjust GGPLOT2 bin size (0 to 1).
 #' @param x.title Title of the x-axis.
@@ -302,58 +302,55 @@ plot_density_distribution_maf <- function(data, maf.group, aes.colour = aes(y = 
 #' @rdname plot_density_distribution_het
 
 plot_density_distribution_het <- function(data, pop.levels, het.group, aes.colour, adjust.bin, x.title){
-  
+
   POP_ID <- NULL
   HET_O <- NULL
   HET_MAX <- NULL
   HET_MIN <- NULL
   VALUE <- NULL
-  
-  if (is.vector(data) == "TRUE") {
-    data <- read_tsv(data, col_names = T)
+
+  if (is.vector(data)) {
+    data <- readr::read_tsv(data, col_names = T)
   } else {
     data = data
   }
-  
+
   data.summary <- data %>%
-    group_by(LOCUS, POP_ID) %>%
-    summarise(
+    dplyr::group_by(LOCUS, POP_ID) %>%
+    dplyr::summarise(
       HET_MEAN = mean(HET_O),
       HET_MAX = max(HET_O),
       HET_MIN = min(HET_O),
       HET_DIFF = HET_MAX - HET_MIN
     ) %>%
-    melt(
-      id.vars = c("LOCUS", "POP_ID"),
-      variable.name = "HET_GROUP", 
-      value.name = "VALUE"
-    )
-  
+    dplyr::group_by(LOCUS, POP_ID) %>%
+    tidyr::gather(data = ., key = HET_GROUP, value = VALUE)
+
   if (missing(pop.levels) == "TRUE") {
     data.summary <- data.summary
   } else {
     data.summary <- data.summary %>%
-      mutate(POP_ID = factor(POP_ID, levels = pop.levels, ordered = T)) %>%
-      arrange(POP_ID)
+      dplyr::mutate(POP_ID = factor(POP_ID, levels = pop.levels, ordered = T)) %>%
+      dplyr::arrange(POP_ID)
   }
-  
-  graph <- ggplot(data.summary, aes(x = VALUE, na.rm = F))+
-    geom_line(aes.colour, stat = "density", adjust = adjust.bin)+ # pop colored
-    #   scale_colour_manual(name ="Sampling sites", values = colour_palette_sites.pink)+
+
+  graph <- ggplot2::ggplot(data.summary, ggplot2::aes(x = VALUE, na.rm = F)) +
+    ggplot2::geom_line(aes.colour, stat = "density", adjust = adjust.bin) + # pop colored
+    #   scale_colour_manual(name ="Sampling sites", values = colour_palette_sites.pink) +
     #     scale_x_continuous(breaks=c(0, 0.05, 0.1, 0.2, 0.5, 1),
-    #                        labels = c("0", "0.05", "0.1", "0.2", "0.5", "1.0"))+
-    labs(x = x.title)+
-    labs(y = "Density of SNP (scaled)")+
-    expand_limits(y = 0)+
-    theme(
-      axis.title.x = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      axis.title.y = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      legend.title = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      legend.text = element_text(size = 12, family = "Helvetica", face = "bold"),
-      strip.text.y = element_text(angle = 0, size = 12, family = "Helvetica", face = "bold"), 
-      strip.text.x = element_text(size = 12, family = "Helvetica", face = "bold")
+    #                        labels = c("0", "0.05", "0.1", "0.2", "0.5", "1.0")) +
+    ggplot2::labs(x = x.title) +
+    ggplot2::labs(y = "Density of SNP (scaled)") +
+    ggplot2::expand_limits(y = 0) +
+    ggplot2::theme(
+      axis.title.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      axis.title.y = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.title = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.text = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      strip.text.y = ggplot2::element_text(angle = 0, size = 12, family = "Helvetica", face = "bold"),
+      strip.text.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold")
     )
-  
+
 }
 
 
@@ -366,49 +363,49 @@ plot_density_distribution_het <- function(data, pop.levels, het.group, aes.colou
 #' @rdname plot_snp_number_loci
 
 plot_snp_number_loci <- function(before.filter.data, after.filter.data) {
-  
+
   GROUP <- NULL
   SNP_N <- NULL
   POP_ID <- NULL
-  
-  
-  if (is.vector(before.filter.data) == "TRUE") {
-    before.filter.data <- read_tsv(before.filter.data, col_names = T)
+
+
+  if (is.vector(before.filter.data)) {
+    before.filter.data <- readr::read_tsv(before.filter.data, col_names = T)
   } else {
     before.filter.data <- before.filter.data
   }
-  
-  if (is.vector(after.filter.data) == "TRUE") {
-    after.filter.data <- read_tsv(after.filter.data, col_names = T)
+
+  if (is.vector(after.filter.data)) {
+    after.filter.data <- readr::read_tsv(after.filter.data, col_names = T)
   } else {
     after.filter.data <- after.filter.data
   }
-  
+
   number.snp.loci <- before.filter.data %>% # Before
-    group_by (LOCUS) %>%
-    summarise(SNP_N = n_distinct(POS)) %>%
-    mutate(GROUP = rep("pre-filters", n())) %>%
-    rbind(
+    dplyr::group_by(LOCUS) %>%
+    dplyr::summarise(SNP_N = dplyr::n_distinct(POS)) %>%
+    dplyr::mutate(GROUP = rep("pre-filters", n())) %>%
+    dplyr::bind_rows(
       after.filter.data %>% # After
-        group_by (LOCUS) %>%
-        summarise(SNP_N = n_distinct(POS)) %>%
-        mutate(GROUP=rep("post-filters", n()))) %>%
-    mutate (
+        dplyr::group_by(LOCUS) %>%
+        dplyr::summarise(SNP_N = dplyr::n_distinct(POS)) %>%
+        dplyr::mutate(GROUP = rep("post-filters", n()))) %>%
+    dplyr::mutate(
       GROUP = factor(GROUP, levels = c("pre-filters", "post-filters"),
-                     ordered = T)
+                     ordered = TRUE)
     )
-  
-  graph <- ggplot(number.snp.loci, aes(factor(SNP_N)))+
-    geom_bar()+
-    labs(x="Number of SNP per haplotypes")+
-    labs(y="Distribution (number)")+
-    facet_wrap(~GROUP, nrow = 1, ncol = 2)+
-    theme(axis.title.x = element_text(size = 12, family = "Helvetica", face = "bold"),
-          axis.title.y = element_text(size = 12, family = "Helvetica", face = "bold"), 
-          legend.title = element_text(size = 12, family = "Helvetica", face = "bold"), 
-          legend.text = element_text(size = 12, family = "Helvetica", face = "bold"), 
-          strip.text.x = element_text(size = 12, family = "Helvetica", face = "bold"))
-  
+
+  graph <- ggplot2::ggplot(number.snp.loci, ggplot2::aes(factor(SNP_N))) +
+    ggplot2::geom_bar() +
+    ggplot2::labs(x = "Number of SNP per haplotypes") +
+    ggplot2::labs(y = "Distribution (number)") +
+    ggplot2::facet_wrap(~GROUP, nrow = 1, ncol = 2) +
+    ggplot2::theme(axis.title.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+          axis.title.y = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+          legend.title = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+          legend.text = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+          strip.text.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"))
+
   graph
 }
 
@@ -426,22 +423,22 @@ plot_snp_number_loci <- function(before.filter.data, after.filter.data) {
 
 
 plot_snp_position_read <- function(data, aes.colour, y.title) {
-  
+
   COL <- NULL
-  
-  
-  ggplot(data, aes(x = COL, na.rm = T))+
-    geom_line((aes.colour), stat = "density", size = 1, adjust = 0.7)+
-    scale_x_continuous(breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 90))+
-    labs(x = "Nucleotide position on the read")+
-    labs(y = y.title)+
-    expand_limits(x = 0)+
-    theme(
-      axis.title.x = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      axis.title.y = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      legend.title = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      legend.text = element_text(size = 12, family = "Helvetica", face = "bold"), 
-      strip.text.x = element_text(size = 12, family = "Helvetica", face = "bold")
+
+
+  ggplot2::ggplot(data, ggplot2::aes(x = COL, na.rm = TRUE)) +
+    ggplot2::geom_line((aes.colour), stat = "density", size = 1, adjust = 0.7) +
+    ggplot2::scale_x_continuous(breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 90)) +
+    ggplot2::labs(x = "Nucleotide position on the read") +
+    ggplot2::labs(y = y.title) +
+    ggplot2::expand_limits(x = 0) +
+    ggplot2::theme(
+      axis.title.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      axis.title.y = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.title = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.text = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      strip.text.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold")
     )
 }
 
@@ -451,9 +448,9 @@ plot_snp_position_read <- function(data, aes.colour, y.title) {
 #' @title Density distribution of diversity (Gene and Haplotypes)
 #' @description GGPLOT2 Density distribution of diversity (Gene and Haplotypes).
 #' @param data The hapstats summary file or object.
-#' @param aes.x GGPLOT2 aesthetics, 
+#' @param aes.x GGPLOT2 aesthetics,
 #' e.g. aes.x = aes(x = GENE_DIVERSITY, na.rm = T).
-#' @param aes.colour GGPLOT2 aesthetics colour, 
+#' @param aes.colour GGPLOT2 aesthetics colour,
 #' e.g. aes.colour = aes(y = ..scaled.., colour = POP_ID).
 #' @param x.title Title of the x-axis.
 #' @param y.title Title of the y-axis.
@@ -461,27 +458,27 @@ plot_snp_position_read <- function(data, aes.colour, y.title) {
 #' @rdname plot_distribution_diversity
 
 plot_distribution_diversity <- function(data, aes.x, aes.colour, x.title, y.title) {
-  
+
   hapstats.summary <- NULL
 
-  
-  ggplot(hapstats.summary, aes.x)+
-    geom_line(aes.colour, stat = "density", adjust = 0.8)+
-    #   scale_colour_manual(name = "Populations", values = colour_palette_sites.pink, breaks = c("BUR", "GRA", "GUL", "LLI", "ANG", "WEI", "HAY", "GOD"))+
-    #   geom_density(aes(fill=POP_ID, color=NA), alpha=0.4)+
-    #   scale_fill_manual(name="Populations", values=colour_palette_sites.pink, breaks = c("BUR", "GRA", "GUL", "LLI", "ANG", "WEI", "HAY", "GOD"))+
-    #   scale_color_manual(name="Populations", values=colour_palette_sites.pink, breaks = c("BUR", "GRA", "GUL", "LLI", "ANG", "WEI", "HAY", "GOD"))+
-    labs(x = x.title)+
-    labs(y = y.title)+
-    expand_limits(y=0)+
-    theme(
+
+  ggplot2::ggplot(hapstats.summary, aes.x) +
+    ggplot2::geom_line(aes.colour, stat = "density", adjust = 0.8) +
+    #   scale_colour_manual(name = "Populations", values = colour_palette_sites.pink, breaks = c("BUR", "GRA", "GUL", "LLI", "ANG", "WEI", "HAY", "GOD")) +
+    #   geom_density(aes(fill=POP_ID, color=NA), alpha=0.4) +
+    #   scale_fill_manual(name="Populations", values=colour_palette_sites.pink, breaks = c("BUR", "GRA", "GUL", "LLI", "ANG", "WEI", "HAY", "GOD")) +
+    #   scale_color_manual(name="Populations", values=colour_palette_sites.pink, breaks = c("BUR", "GRA", "GUL", "LLI", "ANG", "WEI", "HAY", "GOD")) +
+    ggplot2::labs(x = x.title) +
+    ggplot2::labs(y = y.title) +
+    ggplot2::expand_limits(y=0) +
+    ggplot2::theme(
       legend.position = "none",
-      axis.title.x=element_text(size=12, family="Helvetica",face="bold"),
-      axis.title.y=element_text(size=12, family="Helvetica",face="bold"),
-      legend.title=element_text(size=12, family="Helvetica",face="bold"),
-      legend.text=element_text(size=12, family="Helvetica",face="bold"),
-      strip.text.y=element_text(angle=0,size=12, family="Helvetica",face="bold"),
-      strip.text.x=element_text(size=12, family="Helvetica",face="bold")
+      axis.title.x = ggplot2::element_text(size=12, family="Helvetica",face="bold"),
+      axis.title.y = ggplot2::element_text(size=12, family="Helvetica",face="bold"),
+      legend.title = ggplot2::element_text(size=12, family="Helvetica",face="bold"),
+      legend.text = ggplot2::element_text(size=12, family="Helvetica",face="bold"),
+      strip.text.y = ggplot2::element_text(angle=0,size=12, family="Helvetica",face="bold"),
+      strip.text.x = ggplot2::element_text(size=12, family="Helvetica",face="bold")
     )
 }
 
@@ -491,25 +488,25 @@ plot_distribution_diversity <- function(data, aes.x, aes.colour, x.title, y.titl
 #' @title Box plot of the diversity (Gene and Haplotypes)
 #' @description GGPLOT2 Box plot of the diversity (Gene and Haplotypes).
 #' @param data The hapstats summary file or object.
-#' @param aes.x.y The GGPLOT2 aesthetics, 
-#' e.g. aes.x.y = aes(x = factor(POP_ID), y = GENE_DIVERSITY, na.rm = T). 
+#' @param aes.x.y The GGPLOT2 aesthetics,
+#' e.g. aes.x.y = aes(x = factor(POP_ID), y = GENE_DIVERSITY, na.rm = T).
 #' @param y.title Title of the y-axis.
 #' @export
 #' @rdname plot_boxplot_diversity
 plot_boxplot_diversity <- function(data, aes.x.y, y.title) {
-  ggplot(data, aes.x.y)+
-    geom_violin(trim = F)+
-    geom_boxplot(width = 0.1, fill = "black", outlier.colour = NA)+
-    stat_summary(fun.y = "mean", geom = "point", shape = 21, size = 2.5, fill = "white")+
-    labs(x = "Sampling sites")+
-    labs(y = y.title)+
-    theme(
+  ggplot2::ggplot(data, aes.x.y) +
+    ggplot2::geom_violin(trim = F) +
+    ggplot2::geom_boxplot(width = 0.1, fill = "black", outlier.colour = NA) +
+    ggplot2::stat_summary(fun.y = "mean", geom = "point", shape = 21, size = 2.5, fill = "white") +
+    ggplot2::labs(x = "Sampling sites") +
+    ggplot2::labs(y = y.title) +
+    ggplot2::theme(
       legend.position = "none",
-      axis.title.x = element_text(size = 12, family = "Helvetica", face = "bold"),
-      axis.title.y = element_text(size = 12, family = "Helvetica", face = "bold"),
-      legend.title = element_text(size = 12, family = "Helvetica", face = "bold"),
-      legend.text = element_text(size = 12, family = "Helvetica", face = "bold"),
-      strip.text.x = element_text(size = 12, family = "Helvetica", face = "bold")
+      axis.title.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      axis.title.y = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.title = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      legend.text = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold"),
+      strip.text.x = ggplot2::element_text(size = 12, family = "Helvetica", face = "bold")
     )
 }